requirements.txt
torch>=2.0.0
transformers>=4.36.0
sentence-transformers>=2.3.0
datasets>=2.16.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
wandb>=0.16.0
pyyaml>=6.0
tqdm>=4.65.0
accelerate>=0.25.0
einops>=0.7.0
rich>=13.0.0
